requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
qdrant-client==1.15.1
rank-bm25==0.2.2
openai==1.51.0
tqdm==4.66.5
pandas==2.3.3
